[Refactor] move build functions to own submodule
