+++ setup.cfg
@@ -82,7 +82,7 @@ testing =
 
 [options.entry_points]
 console_scripts =
-    build_dataflow = finn.util.build_dataflow:main
+    build_dataflow = finn.builder.build_dataflow:main
 # Add here console scripts like:
 # console_scripts =
 #     script_name = finn.module:function
