+++ src/finn/qnn-data/build_dataflow/build.py
@@ -31,7 +31,8 @@
 # ./run-docker.sh build_custom /path/to/folder
 
 
-import finn.util.build_dataflow as build
+import finn.builder.build_dataflow as build
+import finn.builder.build_dataflow_config as build_cfg
 
 model_name = "tfc_w1a1"
 platform_name = "Pynq-Z1"
@@ -44,14 +45,14 @@ cfg = build.DataflowBuildConfig(
     # folding_config_file="folding_config.json",
     synth_clk_period_ns=10.0,
     board=platform_name,
-    shell_flow_type=build.ShellFlowType.VIVADO_ZYNQ,
+    shell_flow_type=build_cfg.ShellFlowType.VIVADO_ZYNQ,
     generate_outputs=[
-        build.DataflowOutputType.PYNQ_DRIVER,
-        build.DataflowOutputType.STITCHED_IP,
-        build.DataflowOutputType.ESTIMATE_REPORTS,
-        build.DataflowOutputType.OOC_SYNTH,
-        build.DataflowOutputType.BITFILE,
-        build.DataflowOutputType.DEPLOYMENT_PACKAGE,
+        build_cfg.DataflowOutputType.PYNQ_DRIVER,
+        build_cfg.DataflowOutputType.STITCHED_IP,
+        build_cfg.DataflowOutputType.ESTIMATE_REPORTS,
+        build_cfg.DataflowOutputType.OOC_SYNTH,
+        build_cfg.DataflowOutputType.BITFILE,
+        build_cfg.DataflowOutputType.DEPLOYMENT_PACKAGE,
     ],
     save_intermediate_models=True,
 )
